Extract client adapter
