+++ modules/lsp/src/test/scala/playground/lsp/EmptyConfigTests.scala
@@ -3,7 +3,6 @@ package playground.lsp
 import cats.effect.IO
 import cats.effect.kernel.Resource
 import org.eclipse.lsp4j.DidChangeWatchedFilesParams
-import org.eclipse.lsp4j.MessageType
 import playground.lsp.harness.LanguageServerIntegrationTests
 import playground.lsp.harness.TestClient
 import weaver.*
@@ -24,7 +23,7 @@ object EmptyConfigTests extends IOSuite with LanguageServerIntegrationTests {
         assert.same(
           events,
           List(
-            TestClient.MessageLog(MessageType.Info, LanguageClient.NoChangeDetected)
+            TestClient.MessageLog(MessageType.Info, "No change detected, not rebuilding server")
           ),
         )
       }
